+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2
 - Remove parentheses of numericals in author name
 - Remove role name from author name
+- Strip author name
 
 ## 0.0.1
 - 1st release :tada: